+++ src/components/Form/FormField.tsx
@@ -1,7 +1,13 @@
-import { InputLabel } from '../Label/Label';
+import { InputLabel } from '@/components';
+
 import type { FormFieldProps } from './FormField.types';
 
-const FormField = ({ label, htmlFor, required, children }: FormFieldProps) => {
+export const FormField = ({
+  label,
+  htmlFor,
+  required,
+  children,
+}: FormFieldProps) => {
   return (
     <div className="flex flex-col gap-2">
       <InputLabel htmlFor={htmlFor} required={required}>
@@ -11,5 +17,3 @@ const FormField = ({ label, htmlFor, required, children }: FormFieldProps) => {
     </div>
   );
 };
-
-export default FormField;
