+++ src/pages/PromptDetail/_components/BackToListButton.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
 
-import Button from '../../../components/Button/Button';
+import { Button } from '@/components';
 
 const BackToListButton = () => {
   const navigate = useNavigate();
